+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Stretching some circle buttons when system text size is enlarged
 - Cutting long month headers when system text size is enlarged
 - Occasionally seeing the login screen when opening an item in the web viewer
+- Occasionally briefly seeing the bottom selection controls when starting the app
 
 ### Changed