+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - When opening an item in the web viewer, the viewer state (selection, etc.) is reset
 - Renamed Extension store to Extension catalog
+- Made some icons rounded
 
 ## [1.36.0] - 2025-03-31