+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.37.1] - 2025-06-04
 
 ### Fixed
 
@@ -920,4 +920,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [1.37.0]: https://github.com/Radiokot/photoprism-android-client/compare/1.36.0(55)...1.37.0(56)
 
-[Unreleased]: https://github.com/Radiokot/photoprism-android-client/compare/1.37.0(56)...HEAD
+[1.37.1]: https://github.com/Radiokot/photoprism-android-client/compare/1.37.0(56)...1.37.1(57)
+
+[Unreleased]: https://github.com/Radiokot/photoprism-android-client/compare/1.37.1(57)...HEAD